Some more hacking
I think I had a breakthrough figuring out how to do replay. It is more
like a weave starting with the upstream, merging in downstream commits
that are part of the same change (by following replaces pointers) and
then catching the rest of the downstream commits.

This ignores issues that arise when inserting or dropping commits. I
still need to think about those.